requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
flask-cors==4.0.0
qrcode==7.4.2
Pillow==10.4.0
openpyxl==3.1.5
reportlab==4.2.2
matplotlib==3.9.2
Flask-Login==0.6.3
tzdata==2025.1
